Responsivity Update
Corrected line 129 to make the change when the screen has a max width of 414px instead of 412px for iPhone 6-8. 
Reduced the Card Size for screen max width 684px  so the website does not have a side scroll bar for iPhone 6-8 landscape & Android (Pixel 2) landscape.
Changed the Column gap amount for IPad Landscape so the edges of the card don't reach the edge of the screen.